chore: fix circular dependecies
