+++ src/sdk/account/utils/Utils.ts
@@ -27,10 +27,9 @@ import {
 import type { Transport } from "viem"
 import type { Chain } from "viem/chains"
 import {
-  type MultichainSmartAccount,
   isVersionOlder,
   versionIsAtLeast
-} from ".."
+} from "./getVersion"
 import {
   BICONOMY_TOKEN_PAYMASTER,
   MOCK_MULTI_MODULE_ADDRESS,
@@ -39,8 +38,6 @@ import {
   NEXUS_DOMAIN_TYPEHASH,
   NEXUS_DOMAIN_VERSION
 } from "../../account/utils/Constants"
-import type { Instruction } from "../../clients/decorators/mee/getQuote"
-import type { FeeTokenInfo } from "../../clients/decorators/mee/getQuote"
 import { MEEVersion } from "../../constants"
 import { EIP1271Abi } from "../../constants/abi"
 import {
@@ -628,49 +625,6 @@ export function calculateNonceStorageSlot(sender: Address, key = 0n): Hex {
   )
 
   return finalSlot
-}
-
-/**
- * Returns the MEE versions of the orchestrator account
- * on all the chains involved in the quote request.
- *
- * @param account - The multichain smart account
- * @param instructions - The instructions to be executed
- * @param sponsorship - Whether the quote is sponsored
- * @param feeToken - The fee token
- * @returns An array of chain IDs with their corresponding MEE versions as MeeVersionsWithChainId
- * @example
- * ```typescript
- * const meeVersions = getMeeVersionsForQuote(account, instructions, sponsorship, feeToken)
- * ```
- */
-export function getMeeVersionsForQuote(
-  account: MultichainSmartAccount,
-  instructions: Instruction[],
-  sponsorship: boolean,
-  feeToken?: FeeTokenInfo
-): MeeVersionsWithChainId {
-  const usedChains = new Set<number>()
-  for (const op of instructions) {
-    usedChains.add(Number(op.chainId))
-  }
-
-  // For sponsored flow, we can ignore payment chain because orchestrator account
-  // used there will be not user's orchestrator account but sponsorship account.
-  // For non-sponsored flow, the user's orchestrator account will be performing'
-  // the payment userOp, so we need to make sure its MEE version is consistent.
-  if (!sponsorship) {
-    // if sponsorship is false, the feeToken is defined: see GetQuoteParams type
-    usedChains.add(Number(feeToken!.chainId))
-  }
-
-  return Array.from(usedChains, (chainId) => {
-    const deployment = account.deploymentOn(chainId, true)
-    return {
-      version: deployment.version,
-      chainId
-    }
-  }) as MeeVersionsWithChainId
 }
 
 /**
